requirements/develop.txt
-r base.txt

factory-boy==2.11.1
flake8==3.7.7
flake8-isort==2.7.0
ipython==7.4.0
pytest==4.4.0
pytest-cover==3.0.0
pytest-django==3.4.8
